src/requirements.txt
aiosmtplib          ==   2.0.2      
annotated-types     ==   0.7.0      
anyio               ==   4.6.2.post1
asgiref             ==   3.8.1      
Authlib             ==   1.3.2      
bcrypt              ==  4.2.0
blinker             ==   1.8.2
CacheControl        ==   0.14.0
cachetools          ==   5.5.0
certifi             ==   2024.8.30
cffi                ==   1.17.1
charset-normalizer  ==   3.4.0
click               ==   8.1.7
colorama            ==  0.4.6
cryptography        ==   43.0.1
dnspython           ==   2.7.0
email_validator     ==   2.2.0
fastapi             ==   0.115.2
greenlet            == 3.1.0
h11                 ==   0.14.0
idna                ==  3.10
Jinja2              ==   3.1.4
MarkupSafe          ==   2.1.5
motor               ==   3.6.0
msgpack             ==   1.1.0
mysql-connector-python == 9.0.0
oauthlib           ==    3.2.2
passlib             ==   1.7.4
pip                 ==   24.2
pyasn1              ==   0.6.1
pyasn1_modules      ==   0.4.1
pycparser           ==   2.22
pydantic            ==   2.9.2
pydantic_core       ==   2.23.4
pydantic-settings   ==   2.6.1
PyJWT               ==   2.9.0
PyMySQL             ==   1.1.1
python-dateutil     ==   2.9.0.post0
python-dotenv       ==   1.0.1
python-multipart    ==   0.0.12
requests            ==   2.32.3
requests-oauthlib   ==   2.0.0
rsa                 ==   4.9
setuptools         ==    65.5.0
six                ==    1.16.0
sniffio            ==    1.3.1
SQLAlchemy         ==    2.0.34
sqlparse           ==    0.5.1
starlette          ==    0.40.0
typing_extensions  ==    4.12.2
tzdata             ==    2024.1
urllib3            ==    2.2.3
uv                 ==    0.4.25
uvicorn            ==    0.32.0
Werkzeug           ==    3.0.4